Add logger middleware placeholer & move validations to middleware
